restructure email folder to help the email preview work
