+++ emails/assets/crypto.tsx
