+++ emails/assets/discord.tsx
@@ -7,7 +7,7 @@ export default function DiscordIcon() {
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
-      <g clip-path="url(#clip0_40800_13929)">
+      <g clipPath="url(#clip0_40800_13929)">
         <path
           d="M8.39667 9.1665C8.89667 9.1665 9.30167 9.5415 9.2925 9.99984C9.2925 10.4582 8.8975 10.8332 8.39667 10.8332C7.905 10.8332 7.5 10.4582 7.5 9.99984C7.5 9.5415 7.89583 9.1665 8.39667 9.1665ZM11.6033 9.1665C12.1042 9.1665 12.5 9.5415 12.5 9.99984C12.5 10.4582 12.1042 10.8332 11.6033 10.8332C11.1117 10.8332 10.7075 10.4582 10.7075 9.99984C10.7075 9.5415 11.1025 9.1665 11.6033 9.1665ZM15.7425 1.6665C16.7117 1.6665 17.5 2.4715 17.5 3.469V19.1665L15.6575 17.504L14.62 16.524L13.5225 15.4823L13.9775 17.1015H4.2575C3.28833 17.1015 2.5 16.2965 2.5 15.299V3.469C2.5 2.4715 3.28833 1.6665 4.2575 1.6665H15.7417H15.7425ZM12.4342 13.094C14.3283 13.0332 15.0575 11.764 15.0575 11.764C15.0575 8.9465 13.8225 6.66234 13.8225 6.66234C12.5892 5.71817 11.4142 5.744 11.4142 5.744L11.2942 5.884C12.7517 6.339 13.4283 6.99567 13.4283 6.99567C12.6326 6.54733 11.7555 6.26178 10.8483 6.15567C10.2729 6.09067 9.69161 6.09627 9.1175 6.17234C9.06583 6.17234 9.0225 6.1815 8.97167 6.18984C8.67167 6.2165 7.9425 6.32984 7.02583 6.7415C6.70917 6.88984 6.52 6.99567 6.52 6.99567C6.52 6.99567 7.23167 6.304 8.77417 5.849L8.68833 5.744C8.68833 5.744 7.51417 5.71817 6.28 6.66317C6.28 6.66317 5.04583 8.9465 5.04583 11.764C5.04583 11.764 5.76583 13.0323 7.66 13.094C7.66 13.094 7.97667 12.7007 8.235 12.3682C7.14583 12.0348 6.735 11.3348 6.735 11.3348C6.735 11.3348 6.82 11.3965 6.97417 11.484C6.9825 11.4923 6.99083 11.5015 7.00833 11.5098C7.03417 11.5282 7.06 11.5365 7.08583 11.554C7.3 11.6765 7.51417 11.7723 7.71083 11.8515C8.0625 11.9915 8.4825 12.1315 8.97167 12.2282C9.70443 12.3718 10.4578 12.3746 11.1917 12.2365C11.6191 12.1602 12.0361 12.0341 12.4342 11.8607C12.7342 11.7465 13.0683 11.5798 13.42 11.344C13.42 11.344 12.9917 12.0615 11.8683 12.3857C12.1258 12.7182 12.435 13.094 12.435 13.094H12.4342Z"
           className="fill-primary dark:fill-primary-dark"
