+++ emails/assets/facebook.tsx
@@ -7,7 +7,7 @@ export default function FacebookIcon() {
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
-      <g clip-path="url(#clip0_40800_13927)">
+      <g clipPath="url(#clip0_40800_13927)">
         <path
           d="M11.6668 11.2498H13.7502L14.5835 7.9165H11.6668V6.24984C11.6668 5.3915 11.6668 4.58317 13.3335 4.58317H14.5835V1.78317C14.3118 1.74734 13.286 1.6665 12.2027 1.6665C9.94016 1.6665 8.3335 3.04734 8.3335 5.58317V7.9165H5.8335V11.2498H8.3335V18.3332H11.6668V11.2498Z"
           className="fill-primary dark:fill-primary-dark"
