+++ emails/assets/github.tsx
@@ -7,7 +7,7 @@ export default function Github() {
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
-      <g clip-path="url(#clip0_40800_13923)">
+      <g clipPath="url(#clip0_40800_13923)">
         <path
           d="M10.0001 1.6665C5.39592 1.6665 1.66675 5.39567 1.66675 9.99984C1.6658 11.7492 2.21576 13.4545 3.2386 14.8737C4.26143 16.293 5.7052 17.3541 7.36508 17.9065C7.78175 17.979 7.93758 17.729 7.93758 17.5098C7.93758 17.3123 7.92675 16.6565 7.92675 15.9582C5.83342 16.344 5.29175 15.4482 5.12508 14.979C5.03092 14.739 4.62508 13.9998 4.27092 13.8015C3.97925 13.6457 3.56258 13.2598 4.26008 13.2498C4.91675 13.239 5.38508 13.854 5.54175 14.104C6.29175 15.364 7.49008 15.0098 7.96842 14.7915C8.04175 14.2498 8.26008 13.8857 8.50008 13.6773C6.64592 13.469 4.70842 12.7498 4.70842 9.56234C4.70842 8.65567 5.03092 7.9065 5.56258 7.32234C5.47925 7.114 5.18758 6.25984 5.64592 5.114C5.64592 5.114 6.34342 4.89567 7.93758 5.969C8.61597 5.78072 9.31689 5.68596 10.0209 5.68734C10.7292 5.68734 11.4376 5.78067 12.1042 5.96817C13.6976 4.88484 14.3959 5.11484 14.3959 5.11484C14.8542 6.26067 14.5626 7.11484 14.4792 7.32317C15.0101 7.9065 15.3334 8.64567 15.3334 9.56234C15.3334 12.7607 13.3859 13.469 11.5317 13.6773C11.8334 13.9373 12.0942 14.4373 12.0942 15.219C12.0942 16.3332 12.0834 17.229 12.0834 17.5107C12.0834 17.729 12.2401 17.989 12.6567 17.9057C14.3109 17.3472 15.7483 16.284 16.7667 14.8658C17.785 13.4476 18.3329 11.7458 18.3334 9.99984C18.3334 5.39567 14.6042 1.6665 10.0001 1.6665Z"
           className="fill-primary dark:fill-primary-dark"
