+++ emails/assets/index.ts
