+++ emails/assets/iq-logo.tsx
@@ -7,7 +7,7 @@ export default function IQLogo() {
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
-      <g clip-path="url(#clip0_40713_16044)">
+      <g clipPath="url(#clip0_40713_16044)">
         <path
           d="M61.7659 34.5056C61.7659 35.7806 61.5322 37.0003 61.1043 38.1247C59.6404 41.9654 55.921 44.6936 51.5622 44.6936C47.1013 44.6936 43.3106 41.8357 41.9229 37.853C33.682 38.4446 28.9698 34.7666 27.0104 27.7565C22.0609 26.4123 18.4229 21.8936 18.4229 16.525C18.4229 13.053 19.9455 9.93528 22.3604 7.80409C24.4153 5.98755 27.1178 4.88556 30.0785 4.88556C33.4703 4.88556 36.5246 6.33166 38.652 8.6432C40.0829 7.70885 41.7928 7.16726 43.6294 7.16726C44.9111 7.16726 46.1328 7.43209 47.2398 7.90958C50.443 9.29015 52.6952 12.4525 52.7453 16.1463C52.747 16.1895 52.747 16.231 52.747 16.2744L52.7453 16.4284V16.4301C53.2111 16.3279 53.6945 16.2744 54.1919 16.2744C56.4128 16.2744 58.386 17.3366 59.6297 18.9801H59.6317C60.4943 20.1202 61.0055 21.5405 61.0055 23.0803C61.0085 24.6772 60.4451 26.2235 59.4151 27.445C59.3377 27.5393 59.2573 27.631 59.1741 27.7201C60.7872 29.5211 61.7658 31.8999 61.7658 34.5056"
           fill="#FF5CAA"
@@ -15,24 +15,24 @@ export default function IQLogo() {
         <path
           d="M43.5159 15.7014C41.6698 17.7049 40.5424 20.3792 40.5424 23.3167C40.5424 29.5292 45.5858 34.5657 51.8071 34.5657C52.6097 34.5661 53.4101 34.4812 54.1948 34.3125M29.5856 12.0702C29.8674 16.133 33.257 19.3416 37.397 19.3416M49.8498 17.8328C50.687 17.1381 51.6803 16.6562 52.7446 16.4282M38.6513 8.64124C37.34 9.49397 36.2622 10.6758 35.5365 12.0702"
           stroke="#0F172A"
-          stroke-width="1.38598"
-          stroke-miterlimit="10"
-          stroke-linecap="round"
+          strokeWidth="1.38598"
+          strokeMiterlimit="10"
+          strokeLinecap="round"
           stroke-linejoin="round"
         />
         <path
           d="M61.7655 34.5037C61.7655 40.1314 57.1971 44.6919 51.5617 44.6919C47.1008 44.6919 43.3102 41.834 41.9224 37.8513C33.6815 38.4429 28.9693 34.7649 27.0099 27.7547C22.0604 26.4106 18.4224 21.8919 18.4224 16.5233C18.4224 10.0947 23.6403 4.88385 30.078 4.88385C33.4698 4.88385 36.5241 6.32995 38.6515 8.64149C40.0824 7.70714 41.7923 7.16555 43.6289 7.16555C48.6649 7.16555 52.7465 11.2434 52.7465 16.2727L52.7449 16.4267V16.4284C53.2107 16.3262 53.694 16.2727 54.1914 16.2727C57.9541 16.2727 61.005 19.3191 61.005 23.0786C61.005 24.8727 60.3102 26.5024 59.1736 27.7184C60.7867 29.5194 61.7653 31.8981 61.7653 34.5038L61.7655 34.5037Z"
           stroke="#0F172A"
-          stroke-width="1.21265"
-          stroke-miterlimit="10"
+          strokeWidth="1.21265"
+          strokeMiterlimit="10"
           stroke-linejoin="round"
         />
         <path
           d="M59.1738 27.7182C59.1738 27.7182 58.3828 26.8985 57.3008 26.0454M28.7474 46.1951C28.3008 45.2635 28.0151 44.2632 27.9022 43.2365"
           stroke="#0F172A"
-          stroke-width="1.38598"
-          stroke-miterlimit="10"
-          stroke-linecap="round"
+          strokeWidth="1.38598"
+          strokeMiterlimit="10"
+          strokeLinecap="round"
           stroke-linejoin="round"
         />
         <path
@@ -42,39 +42,39 @@ export default function IQLogo() {
         <path
           d="M45.2746 34.5042C46.8734 35.6839 48.0967 37.2792 48.755 39.1095M15.6746 41.246C15.4528 40.4606 15.3349 39.6319 15.3349 38.7738C15.3349 35.272 17.3082 32.243 20.1737 30.7949M12.4437 22.9457C11.5949 23.0322 10.7806 23.2312 10.0148 23.5304M27.0999 16.5237L27.098 16.522M27.098 16.522C25.8803 17.1794 24.8235 18.0758 23.9972 19.1449"
           stroke="#0F172A"
-          stroke-width="1.38598"
-          stroke-miterlimit="10"
-          stroke-linecap="round"
+          strokeWidth="1.38598"
+          strokeMiterlimit="10"
+          strokeLinecap="round"
           stroke-linejoin="round"
         />
         <path
           d="M48.7554 39.1091C49.1161 40.1084 49.3001 41.1628 49.2992 42.225C49.2992 47.6885 44.4971 52.1155 38.5721 52.1155C34.1788 52.1155 30.4037 49.6813 28.7476 46.1954C27.3826 47.121 25.7452 47.6607 23.9869 47.6607C20.0441 47.6607 16.7178 44.9515 15.6749 41.2458C14.9304 41.4311 14.1659 41.5247 13.3985 41.5243C8.24807 41.5243 4.07117 37.3547 4.07117 32.2096C4.07117 28.2583 6.53635 24.8813 10.015 23.5301C9.87818 22.9645 9.80542 22.378 9.80542 21.7759C9.80542 17.11 14.1469 13.3281 19.5035 13.3281C22.5786 13.3281 25.3225 14.577 27.0982 16.5216C28.5224 15.7587 30.1647 15.3211 31.9145 15.3211C37.3421 15.3211 41.7422 19.5234 41.7422 24.7065C41.7428 24.8288 41.7399 24.951 41.7335 25.0731C46.212 25.8931 49.6161 30.0005 49.6161 34.9448C49.6161 36.4309 49.3096 37.8409 48.7553 39.109L48.7554 39.1091Z"
           stroke="#0F172A"
-          stroke-width="1.21265"
-          stroke-miterlimit="10"
+          strokeWidth="1.21265"
+          strokeMiterlimit="10"
           stroke-linejoin="round"
         />
         <path
           d="M33.3507 27.8634C35.0762 26.0504 37.4582 24.9275 40.0879 24.9275C40.6493 24.9275 41.1985 24.9777 41.7337 25.0761M41.1167 36.8607C39.7572 38.5583 37.6657 39.6458 35.3198 39.6458C31.2228 39.6458 27.9016 36.3289 27.9016 32.2378C27.899 30.5295 28.4901 28.8732 29.5741 27.5516"
           stroke="#0F172A"
-          stroke-width="1.38598"
-          stroke-miterlimit="10"
-          stroke-linecap="round"
+          strokeWidth="1.38598"
+          strokeMiterlimit="10"
+          strokeLinecap="round"
           stroke-linejoin="round"
         />
         <path
           d="M38.6002 43.3122C38.6002 41.8573 38.2327 40.4891 37.5847 39.2935M23.0966 38.5808C24.3197 36.7711 26.2238 35.4614 28.443 35.0169M9.87158 32.2175C10.8113 31.8457 11.8129 31.655 12.8236 31.6552C14.4442 31.653 16.027 32.1444 17.3607 33.0637M25.1229 25.3995C24.6181 25.5479 24.0838 25.6274 23.5309 25.6274C22.0632 25.6294 20.6531 25.0573 19.6026 24.0337M49.4693 25.0382C50.2458 24.4597 51.189 24.1478 52.1578 24.1493C52.4271 24.1491 52.6958 24.173 52.9608 24.2205"
           stroke="#0F172A"
-          stroke-width="1.38598"
-          stroke-miterlimit="10"
-          stroke-linecap="round"
+          strokeWidth="1.38598"
+          strokeMiterlimit="10"
+          strokeLinecap="round"
           stroke-linejoin="round"
         />
         <path
           d="M59.9728 27.7227C61.0331 26.4089 61.6114 24.7799 61.6114 23.0791C61.6114 18.9924 58.2827 15.6675 54.1914 15.6675C53.9054 15.6675 53.6201 15.6841 53.3371 15.7168C53.0476 10.6191 48.8037 6.56064 43.6287 6.56064C41.9032 6.56064 40.2271 7.01179 38.7504 7.86972C36.4494 5.57356 33.3306 4.28222 30.0777 4.27881C24.6257 4.27881 19.9955 7.85156 18.406 12.7749C13.2383 13.2577 9.19885 17.1101 9.19885 21.7764C9.19885 22.2408 9.23951 22.7059 9.31968 23.1647C5.74579 24.7643 3.46472 28.2608 3.46472 32.2102C3.46472 37.6801 7.92097 42.1302 13.3984 42.1302C14.0239 42.1303 14.648 42.0719 15.2626 41.9559C16.5627 45.6991 20.0727 48.2666 23.9868 48.2666C25.575 48.2677 27.1355 47.8507 28.5109 47.0577C30.4474 50.5195 34.321 52.7214 38.5719 52.7214C43.7473 52.7214 48.1205 49.4907 49.4717 45.0935C50.1513 45.2273 50.8504 45.2977 51.5615 45.2977C57.5222 45.2977 62.3718 40.4555 62.3718 34.5041C62.3718 32.0248 61.5236 29.6369 59.9728 27.7226V27.7227Z"
           stroke="#0F172A"
-          stroke-width="2.42543"
-          stroke-miterlimit="10"
+          strokeWidth="2.42543"
+          strokeMiterlimit="10"
           stroke-linejoin="round"
         />
         <path
