+++ emails/assets/nfts.tsx
