+++ emails/assets/organization.tsx
