+++ emails/assets/people-of-crypto.tsx
