+++ emails/components/banner.tsx
