+++ emails/components/content.tsx
