+++ emails/components/footer.tsx
