+++ package.json
@@ -30,7 +30,8 @@
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "test:e2e": "jest --config ./test/jest-e2e.json"
+    "test:e2e": "jest --config ./test/jest-e2e.json",
+    "email:dev": "email dev -p 3001"
   },
   "dependencies": {
     "@apollo/server": "^4.10.4",
@@ -64,6 +65,7 @@
     "@types/express-serve-static-core": "^4.17.35",
     "@types/got": "^9.6.12",
     "@types/nodemailer": "^6.4.6",
+    "@types/react-dom": "^19.0.4",
     "@types/remove-markdown": "^0.3.1",
     "apollo-server-express": "^3.13.0",
     "apollo-server-fastify": "^3.13.0",
@@ -98,6 +100,8 @@
     "pg": "^8.7.3",
     "pg-boss": "^8.1.1",
     "pm2": "^5.2.2",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "reflect-metadata": "^0.1.13",
     "remove-markdown": "^0.5.0",
     "rimraf": "^6.0.1",
@@ -157,6 +161,10 @@
     "ts-node": "^10.9.2",
     "tsconfig-paths": "^3.11.0",
     "typescript": "^4.6.2"
+  },
+  "resolutions": {
+    "wrap-ansi": "7.0.0",
+    "string-width": "4.1.0"
   },
   "jest": {
     "moduleFileExtensions": [
